Update installer.sh
fixed download stuff, not finished.

+++ installer.sh
@@ -6,7 +6,7 @@ function animate_dots() {
   local char="."
   while [[ $count -lt 5 ]]; do
     echo -en "\r[Macware]: $1$char"  # Move cursor to beginning of line (-en)
-    sleep 0.2
+    sleep 3
     char="$char."
     ((count++))
   done
@@ -20,17 +20,30 @@ animate_dots "Grabbing resources..."
 animate_dots "Downloading Macware.."
 animate_dots "Installing latest Roblox application"
 
-echo "Download complete!"
+# Prompt user for confirmation
+echo -e "\n[Macware]: This script will uninstall/reinstall your roblox."
+echo -e "[Macware]: Continue? (y/N)"
 
-echo -e "Downloading Latest Roblox..."
+read -r -p "" response
+
+# Check user response (case-insensitive)
+if [[ $response =~ ^[Yy]$ ]]; then
+  # User wants to continue 
+  echo -e "[Macware]: Grabbing the latest roblox version."
     [ -f ./RobloxPlayer.zip ] && rm ./RobloxPlayer.zip
     local version=$(curl -s "https://clientsettingscdn.roblox.com/v2/client-version/MacPlayer" | ./jq -r ".clientVersionUpload")
     curl "http://setup.rbxcdn.com/mac/$version-RobloxPlayer.zip" -o "./RobloxPlayer.zip"
     rm ./jq
 
-    echo -n "Installing Latest Roblox... "
+  animate_dots -n "Installing Latest Version of Roblox (don't close, will take time.)"
     [ -d "/Applications/Roblox.app" ] && rm -rf "/Applications/Roblox.app"
     unzip -o -q "./RobloxPlayer.zip"
     mv ./RobloxPlayer.app /Applications/Roblox.app
     rm ./RobloxPlayer.zip
     echo -e "Done."
+
+else
+  # User doesn't want to continue
+  echo "[Macware]: Installation cancelled."
+fi
+
